Changed values in map directly to reflect 1.0 pbr, instead of dividin…

+++ SlotsProto/CombinationEvaluator.h
@@ -6,8 +6,8 @@ class CombinationEvaluator
 public:
    CombinationEvaluator();
    CombinationEvaluator(int);
-   size_t checkPayLineCombinations(const std::vector<SymbolCombination>&);
+   double checkPayLineCombinations(const std::vector<SymbolCombination>&);
 private:
-   std::map<SymbolCombination, int> evalmap; //Should change bool to int/double(amount you would win)
+   std::map<SymbolCombination, double> evalmap; //Should change bool to int/double(amount you would win)
 };
